pyproject.toml
[project]
name = "photokiller"
version = "0.1.0"
description = "Lean photobooth app for Raspberry Pi with DSLR support, CUPS printing, and PySide6 UI"
requires-python = ">=3.10"
authors = [
  { name = "Photokiller" }
]
dependencies = [
  "PySide6>=6.6",
  "Pillow>=10.3",
  "pydantic>=2.7",
  "pycups>=2.0.1",
  "opencv-python>=4.9; platform_machine != 'armv7l'",
  "opencv-python-headless>=4.9; platform_machine == 'armv7l'",
  "typing-extensions>=4.11",
]

[project.scripts]
photokiller = "app.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100

[tool.pyinstaller]
name = "photokiller"


